Don't put the string of the expression in the round() exception
This fixes the specific issue from #8309, but there are likely more cases that
should be fixed as well. This leads to great speed improvements for large
expressions in any function that tries to call int() on it to see if it is an
integer.

In the future, we may want to keep the expression on the exception object as
well, like on a .expr attribute of the exception object, but I have not done
this for now, as we should decide on a uniform way to approach this problem.