remove _fuzzy_group_inverse; add notes to docstring
_fuzzy_group_inverse (except with the quick_exit option)
is just fuzzy_and. The quick_exit option is never used with
_fuzzy_group_inverse and the presence of this function leads
to confusion about which function to use and with which option.

It would still be nice to get a more descriptive keyword for this
function. A few ideas:

FF=None (with `FF=not None` the default).

    The nice thing about that is its self documenting nature: two
    Falses should return as None. The bad thing is that using None
    for a keyword to make it do something is kind of odd.

permit_one_false=True (default False).

    That doesn't tell what happens when there is a second False.

Other ideas...

second_False_gives_None=True
strict=True