Fix vector printing tests in ASCII
I actually didn't fix them, but the tests pass. I had to punt on this one,
because the printer uses some awful hacks, and I don't really know how it is
supposed to work. So now vector code will raise NotImplementedError when ASCII
pretty printed.

Also clean up a bit of bad code that was reassigning self and made it slightly
less confusing.