solve modifications
o move unrad to fallback

o keep only bases of denoms

  Although one might want the Pow form of a denominator when trying
  to find the domain of a function, it is not needed for zero testing.

o move hyperbolic rewriting out of _solve

o simplify which denominators are checked; checking solutions is time
  consuming, so if the number of expressions to check is reduced this
  will help keep solve from bogging down in that step. e.g. if x**3 and
  x appear as denominators, only the x needs to be checked.

o solve equations in the form 0**f(x) - 1

o always obtain a solution to cubics, etc... (issue 8755)

note:

The former solution to the expression in the test for issue 6644
came about by applying unrad to the eq to remove all radicals. This
produced a result that factored into a monomial, a quadratic and
an expression of degree 6. The monomial solution (0) is not valid
and only one of the quadratic solutions was valid. All 4 solutions
are now returned. For speed purposes, the spurious root is included
when checking is disabled.