Float: losslessly convert strings to high-precision Floats
Previously, the output of `sympify(<str>)` always gave precision
of 15, even if the string was much longer.  `N(<str>)` and Float
were similar.  For Float, a user could always to `Float(str, '')`
to auto select the precision.  This is good but perhaps not ideal
for `sympify(str)` (which ultimately calls `Float(str)` without
the nullstring.

The new behaviour is all of sympify, N, and Float will now default
to a minimum precision of 15 when converting from a string (thus
preserving previous behaviour).  But when given a long string
they can use additional precision automatically.

This is accomplished with a new choice of the `prec` kwarg in the
Float constructor: `prec='min15'` uses a minimum precision of 15
but can use higher for strings, where is behaves like the null
string.  This new choice is made the default.

Fixes #8821.