Float: make some tests stricter when comparing two Floats
For `==` and `basic._aresame` are rather weak tests of two Floats
being equal.  Both are False if args have different precisions.
In these tests, we want more than that.  We make sure they are
structurally equal (`==`) and that they have the same `_prec`.
An example for insufficiency of `_aresame` is:
````
p = Float('1.23', '')
q = Float(p, 15)
_aresame(p, q)   # True
````
(maybe sometimes those should be considered equal, but not in
these tests).

We add a new `same_and_same_prec` function (local to the test file).