watch for all rationals in calc_part
The number reresented with n, c, p in calc_part is [1, -1][n]*c*2**p.
with n,c,p all integers. A non-integer will have p < 0 so that test was
added to the existing test. Th c!=1 and p!=0 test is for a computed
integer; something that was a pure integer would have c=1 or p=0.