Fix _print_AlgebraicNumber
The code for _print_AlgebraicNumber(self, expr) has `self` in
two places where `expr` is required. Moreover, the order of the
arguments in the printed representation is inverted in relation
to __new__(cls, expr, coeffs=Tuple(), alias=None, **args)
in AlgebraicNumber.

It is also necessary to set scoeffs along with rep in __new__ when
coeffs == () in order to satisfy eval(srepr(expr)) == expr.
Some tests have to be modified accordingly.