update handling of systems of equations
Better attention is paid to checking solutions: linear systems are not
checked since they come from routines that are already tested. (But
denominators are checked). Also, checking of solutions as they are found
is not done since all the checking is done at the end.

Also, only do simplification of results not obtained from
system-of-equation solvers unless an explicit 'simplify' flag is set.