Make _rewrite_gamma() raise IntegralTransformError (fixes #8882)
inverse_mellin_transform() applies _rewrite_gamma() to check whether
the transform can be expressed in terms of Meijer G-functions.
A failure is indicated by raising IntegralTransformError.

For the rewrite it is necessary that the arguments of trigonometric
and gamma functions are of the form `a*pi*s + b` or `a*s + b`,
respectively, where all the multipliers `a` are rational multiples of
a single real number (which can be absorbed into `s`).

The code is modified in phase 1) to test this for all multipliers
instead of only manifestly real ones.