8907: improve comp for testing for (in)significant difference
A test as simple as Mod(1.3, 1) == .3 fails in SymPy because the
underlying Floats that end up being compared are not exactly the
same. To fascillitate such comparisons, changes were made to the
comp function. To compare such numbers the comp function can be
used as comp(Mod(1.3, 1), .3) and the result will be True.