Add table and cache for derivatives of Bessel functions to heurisch
The standard forms of derivatives of Bessel functions of order ``n``
involve functions of neighboring orders ``n - 1`` and ``n + 1``.
They cannot be used as such in heurisch because iterated derivation
will introduce new orders ``n - 2`` and ``n + 2``.

It is possible to reduce the number of new orders by means of the
recurrence relation between Bessel functions of orders ``n - 1``,
``n`` and ``n + 1``. Thus the derivative can be expressed in terms
of the function of order either ``n - 1`` or ``n + 1`` along with
the original function of order ``n``.

Each one of these alternative expressions, if used exclusively, has
the same flaw as the standard form. The solution is to use both forms
to express the derivatives of the functions of orders ``n - 1`` and
``n`` in terms of the two of them. To make certain that the appropriate
form is used for each order, both expressions are cached as soon as
one is encountered.