issue #8972: patches for two failures in pypy
There are two recurring failures that are caused by different
implementations of dictionaries in cpython and pypy.

almosteq() in rings.py is changed to compare two polynomials by
looping over keys even if their orders differ.

The example of _remove_gens() in combinatorics/utils.py is changed
by not checking the result of the Schreier-Sims algorithm which
depends on the implementation fo dictionaries.
This is harmless since only the result of removing redundant
generators is interesting.

Signed-off-by: Aaron Meurer <asmeurer@gmail.com>