rootof: fix numerical evaluation of RootOf
RootOf._eval_evalf() hangs in the case when the computed root lies on
the boundary of the interval containing it. This occurs even in the
following simple example.

In [1]: r = RootOf(x**2 + 1, 0, radicals=False)
In [2]: r.evalf()   # hangs indefinitely

The solution is to allow for such boundary interval values.
Fixes #9019. The above example is included as a test.