pretty printing: improved `Trace`
Defined a function _print_Trace in sympy.printing.pretty.pretty.PrettyPrinter
so that this function is used to print the Trace rather than the
emptyPrinter.

Example:

Before:

In [70]: X = ImmutableMatrix([[1, 2], [3, 4]])

In [71]: w = Trace(X) + Trace(2*X)

In [72]: pprint(w)
Trace(Matrix([ + Trace(Matrix([
   [1, 2],          [2, 4],
  [3, 4]]))        [6, 8]]))

After:

In [70]: X = ImmutableMatrix([[1, 2], [3, 4]])

In [71]: w = Trace(X) + Trace(2*X)

In [72]: pprint(w)
  ⎛⎡1  2⎤⎞ +   ⎛⎡2  4⎤⎞
tr⎜⎢    ⎥⎟   tr⎜⎢    ⎥⎟
  ⎝⎣3  4⎦⎠     ⎝⎣6  8⎦⎠

The position of the operators in between the Traces are off position.
Still need to figure out how to fix that.