Operator: __mul__ exception for identity mul
The multiplication of operator with identity returns the operator.
Example:
In [2]: A = HermitianOperator('A')

In [3]: IdentityOperator()*A
Out[3]: A

In [4]: A*IdentityOperator()
Out[4]: A

In [5]: B = HermitianOperator('B')

In [6]: A*B
Out[6]: A⋅B

test_opertor.py: tests updated

A test added to check whether A*IdentityOperator() == A
where A is a operator. Issue #9056