Deleted unreachable code in class jacobi in polynomial.py
The deleted condition a == -b is the exactly the same as the its above if else ccondition b == -a. b == -a condition check whether or not two symbols are opposite so is a == -b. Since b == -a is a condition before a == -b, then a == -b condition is never reached. Thus, that condition statement is unreachable.