Fix Pow._eval_subs giving wrong results when no-cache
Just a simple fix, change an 'is' to a '=='.

Example: (1/sqrt(-x**2+16)).subs(sqrt(-x**2+16), y) didn't do the
substitution when SYMPY_USE_CACHE='no'. It worked with the cache on,
however.

I've added a test for the above example too. I can approximate turning
the cache off by clearing the cache. The test didn't work before this
change and now it does.