More work on Transform for Point/Point3D
Corrected the transform code for Point, which was attempting to
transform by the transpose of the matrix supplied. Now it applies the
correct matrix. Added type checking to ensure that the correct input is
supplied to transform for both Point and Point3D. Removed some
incorrect tests for Point transform.