rootofsimp: handle certain exceptions
Throw a meaningful error message when the denominator of an expression
containing RootOfs is not invertible modulo its defining polynomial.
In the case when the expression is not polynomial in the RootOf
skip any simplification attempts. This behavior can be improved if
polynomial subexpression detection is used to simplify arguments of
non-polynomial functions containing RootOfs.