Fix Mul._eval_is_prime infinite recursion when no cache
An infinite recursion can be triggered in Mul._eval_is_prime because this function calls r.is_prime. r can be equal to self. The recursion only appears when SYMPY_USE_CACHE='no' because the caching system will return the assumption as None when recursing on itself.

To see the issue, run:
./bin/test -C test_assumptions -k test_issue_4822