Fix Pow._eval_is_prime infinite recursion when no cache
An infinite recursion can be triggered in Pow._eval_is_prime because this function calls self.doit().is_prime. self.doit() can be the same as self. The recursion only appears when SYMPY_USE_CACHE='no' because the caching system will return the assumption as None when recursing on itself.

To see the issue, run:
./bin/test -C test_assumptions -k special_assumptions