Prevent import cycles via assumptions.ask
ask.py would collect known facts during import, calling back into SymPy.
Replaced the statements with cached functions that would do it, so the
collection passes will happen only after SymPy is fully loaded.

The exec-based consistency test would not work for a function-based
facts collection; for some reason, it would not grok the import of
Or (and most likely other symbols).
I suspect that the global namespace is lost somewhere along the way
from exec String to cache to function, but I could not find that out;
instead, I resorted to a textual compare of ask_generated.py with what
compute_known_facts() returns.
If the consistency test is just to prevent an outdated ask_generated.py,
then that works, too. If the original consistency test prevented other
mishaps not tested for elsewhere, then that needs fixing.