Fix value in __wrapped__ in decorator.
Some versions of update_wrapper erroneously assign the final
function of the wrapper chain to __wrapped__, see
https://bugs.python.org/issue17482 .

This is relevant only if multiple decorators are applied to the same
function/class, and if an outer decorator needs to do introspection.
We do not have that yet, but it's a use case for the upcoming `@lazy`
decorator which is supposed to be combinable with `@property` and
`@staticmethod`.