Unifying future imports
Here, making the `from __future__ import` lines adhere to a common
standard, namely:

1) Always the same in all pertinent files, so no discussion is required
whether it's needed in a particular file or not. Just boilerplate that
can be copied over and be done with it.

2) Individual items on the import are sorted alphabetically. The
actual order isn't that relevant, but having a fixed order helps
avoid needless discussion, and alphabetical is as good as any.

3) Files in code copied from other projects were not touched. These
follow different rules than SymPy's conventions.
I think this applies to the files in build/ and doc/.
I didn't touch release/ either because it's just a deployment script,
no SymPy code; maybe I should have, just for consistency.
I also left the empty `__init__.py` files as they are.

Since I have been accused of being the master of bikeshedding, I lost
interest in pursuing this further. What remains to be done are the
subdirectories in sympy/ that are alphabetically after matrices/ ;
feel free to take over.