Fix bug in gaunt: it should now vanish when sum of l's is odd
The Gaunt coefficient must vanish if l1 + l2 + l3 is odd.  This wasn't
the case previously (#7586) due to a buggy condition check.  A
regression test has been added for this.

In addition, a series of tests to ensure the consistency of Gaunt
coefficients and Wigner 3j symbols have been added.  Remark: these tests
are kind of slow (~10 sec).