Add ring series docs as a separate file into ringseries.rst
The ring series module does not fit in the internals of polys as it does not
adhere to the basic definition of a polynomial. Moreover, its application is
quite different from the core polynomials.

Since the goal is to make the ring series full fledged it should be separated
from the internals of polynomials.