solvers: Added to the solve() docstring
Working the final example in the tutorial when solving for the radius of a
sphere in terms of volume I could not figure out how to alleviate unwanted
solutions.  Being a Mathematica man I assumed the assumptions would be an
argument to solve() but there was little mention of it in the docstring.
I myself feel that this should at least be mentioned, so I added it to the
note.