Fixed python3 compatibility and removed print in evaluate method
the module used xrange() at some points. I changed these to the range
funciton from sympy.core.compatibility

There was a problem with lcm(map( .. )) in python 3. The lcm() function
(least common multiple) can work with a list but not a generator object,
which map() returns in python 3.
I fixed this issue by using lcm(list(map( .. )))

I removed the printing of timesteps in the evaluate method.

The tests now all pass on my machine in python 2.7 and python 3.4