IPython printing defaults
In SymPy there are currently implemented IPython output formatters
for three MIME types.

- image/png
- text/latex
- text/plain

There are two types of image/png formatters.

- latex/dvipng
- matplotlib

Since matplotlib.mathtext cannot handle LaTeX environments, the latex/dvipng
combination should be preferred.

When running a code cell, IPython will apply each one of the configured
formatters to the code, and send all the results to the frontend
(qtconsole or notebook) excluding 'None' results.  If IPython is run from
an ordinary terminal, only the text/plain formatter is applied.

Frontends are at liberty to choose any of the formatted results.

qtconsole will prefer image/png, and fall back to text/latex if that is
unavailable, and finally to text/plain if there is nothing else.
It will format text/latex by using matplotlib, and fail if that is
not available or there are LaTeX environments. Hence the text/latex formatter
should not be configured when there is no matplotlib or it is used for
image/png. (If it failed when formatting image/png, it would fail again.)

notebook does not make any choices but will forward all available output
data to the terminal browser. It does base64 encode the png image for the
transfer but latex and plain texts are sent unprocessed. In addition,
notebook will send the URL of a MathJax installation, remote or local,
if available. How that is used will depend on the browser configuration.

These observations suggest the following printing default settings in
SymPy.

- If latex and dvipng executables are available, set use_latex=True
to configure all three formatters using latex/dvipng for image/png.
- Else, if matplotlib is installed, set use_latex='matplotlib' to
configure the formatter for image/png (and text/plain) but not for text/latex.
- Else set use_latex=False to turn off all LaTeX processing leaving
only text/plain.

These settings are needed to suppress errors in qtconsole under
different conditions. The use_latex='matplotlib' condition is unnecessarily
restrictive for notebook if MathJax is available, but that can still be
activated by setting use_latex='mathjax' in init_printing.