9451: rework inequality solving
Issues 9451 and 9954 stem from some problems with how inequalities
functions are used.

The primary entry point is reduce_inequalities. This function should
dispatch to the appropiate helper.

-- issue 9954 stemmed from the fact that the change to a real variable
   was happening in solve_univariate_inequality. This has been removed.

-- issue 9451 stemmed from the fact that the expression was dispatched
   to reduce_abs_inequalities rather than solve_univariate_inequality.

Changes were also made to reduce_inequalities to make sure that errors
are raised only for non-real symbols of interest. Any rational that
auto-evaluates to a bool is handled there, too, rather than in helper
functions.