Shortcut cancelling fractions with integer one
Fractions with integer one in either the numerator or the denominator
would make it possible to cancelled out any common factor from the
numerator and the denominator.  However, in the current implementation,
there is no special treatment for this situation.  The computation of
the GCD with integer one would take a long time by using the Euclidean
algorithm, as in the issue #9341.

As a result, in this commit, when called with expressions, the cancel
function will directly return the quotient when an integer unity is
found.  This could solve issue #9314.  And when the function is called
with polynomials directly or called with tuples, this short-cut will not
be invoked to better preserve its original behaviour in internal use.