unit tests: for new coordinate system classes
Here are the initial unit tests.  The goal here is to get the classes
back to where I had them in a previous PR (with better naming) while
adding a callable superclass.

The superclass will define certain properties of the system created
such as differential operations, coordinate relations and the metric.
If it is inherited by a subclass new methods will be available such as
rotation_matrix, locate_new and orient_new.