Add one as special case in computing GCD of DMP
In the GCD computation of dense multivariate polynomials, previously in
function `sympy.polys.euclidtools._dmp_rr_trivial_gcd`, only the special
cases with one of the polynomials being zero is specially treated as
trivial case.  However, the polynomial of constant unity is the unit of
the multiplication operation in the ring, so it cannot possibly have
other GCD with any polynomial.  So now one is being handled as a special
case as well to skip the later expensive operations.