Fix latex printing tests of Jz in the quantum module
This is a slight regression, in that Jz**2 now prints as J_z^{2} instead of
\left(J_z\right)^{2}, which doesn't match the other pretty printers. However,
due to the way precedence works in the printers, there is only one precedence,
and making it print with precedence that parenthesizes it also causes it to be
parenthesized for the string printer, which doesn't need it.  I chose to go
with this behavior because didn't require changing any of the quantum
code. The other way can be achieved by adding

    from sympy.printing.precedence import PRECEDENCE
    precedence = PRECEDENCE[Pow']

to JxOp (and similar classes).

A better fix would be to refactor the printers so that precedence can be set
separately for different printers.