modify solve_univariate_inequality and _solve_inequality
Since _solve_inequality is used in several places and it
only solves for linear inequalities, it is now exposed as
a public function.

Since solve_univariate_inequality may be called directly
(though reduce_inequalities is the preferred function to call)
it now calls to solve_linear_inequalities and handles those
as quick-exit values.