apart(): don't go inside noncommutative expressions
This is a slight regression. We need to do this because noncommutative
expressions might not be Expr, which the rest of the routine requires
(as_numer_denom()). In particular, this breaks with MatrixExpr (it currently
works because it is set as iterable(), but that is changing in a future
commit). This code was originally added in #1925.