sympy.core.compatibility.iterable checks the _iterable attribute
Previously (and still currently) classes could allow themselves to be not
iterable according to this function by subclassing from NotIterable. However,
this does not allow subclasses of such classes that should be iterable
override this behavior.

The motivating use-case here is ImmutableMatrix, which subclasses from
MatrixExpr.  MatrixExpr should not be iterable, but ImmutableMatrix should
be. A future commit will set this attribute on those classes.