Set iterable(MatrixExpr) to False
However, iterable(ImmutableMatrix) remains True, even though it is a subclass
of MatrixExpr. Any other subclass that represents an explicit matrix should
also set this. I don't think it would be a good idea to set something like
iterable(MatrixSymbol('x', 2, 2)) to True, even though that might be
considered "explicit", because the iterates of it are not actually its .args.
In general, functions in SymPy will use iterable() to automatically vectorize
over arguments, so this should only be set to True if that behavior is
desired.

This change is needed to make cse() work correctly on matrix expressions (no
tests for that are added yet because some additional changes are also required
to make this work).

This also fixes a test failure in count_ops, which was counting operations
incorrectly for MatrixSymbol, since it thought it was an iterable
(MatrixSymbol is now counted as a single operation).