10268: make interval pt significantly different
There were two problems:

1) the pt being tested was too close to the interval
pt and was thus the difference was not distinguished
from zero.

2) the routine 'is_comparable' wasn't set up to handle
numbers that have already been computed; one could insist
that the input NOT be a Number but for now, a Number is
now handled by not re-evaluating it, but simply taking the
value given. A test that failed in master was added: the
troublesome trig identity is 0 but does not evaluate with
precision so it is not comparable (hence, an inequality
involving its comparison to 0 should not be evaluated)

>>> n = cos(1)**2 + sin(1)**2 - 1
>>> n < 0
-1 + cos(1)**2 + sin(1)**2 < 0