unwrap nested And/Or; keep only extremum
Although unwrapping occurs in the LatticeOp class,
unless And and Or args are unwrapped in the And and
Or class they won't enter the simplification logic
there.

Or and And are now fully aware of extremum so
And(x>1, x>2 x<0) is false while Or(x<1, x<2, x>0) is true.