modular: Calculating Modular Multiplicative Inverse
Given two number a  and m.This function computes the number
c such that   (a*c)%m==1.
Meaning that if a is multiplied by c then the remainder obtained
on division with m is 1.  The value of m need not be a prime.
This function  returns  the value of c  if some c exists.
Otherwies it raises an Exception stating that modular inverse does not
exist.

The time taken for running  hundred thousand iterations of
the same input varied from .8 to 3.6 seconds depending on the
input,on a 4gb ram,i3 processor.

Examples

modular_inverse(3,11)=4
Because 3*4=12   and 12 = 1 mod(11)

modular_inverse(2,5)=3
Because 2*3=6   and 6 = 1 mod(5)