10323: correct get_integer_part logic
The conditions on when to do the calculation or not
were there to keep known integer expressions from
chasing to PrecisionExhaustion as they tried to evaluate.
p==0 is the only condition that indicates an expression
that is already known to be an integer.  p<0 would indicate
a rational (hence not exact and in need of evaluation) but
there are other expressions that will have p>0 that need
the evaluation that was being guarded by the conditions.
So now only the p != 0 condition remains but substutions
(which seem to be the source of the problems in this
portion of code) are now made if they involve numbers
with integral re and im parts. (It's a little ugly because
a substitution like {x: 3} has an <int> type, not Integer.)