[WIP]`SymmetricGroup` is now a class not function
Here is what @jksuom (https://github.com/jksuom) mentioned about the current
`SymmetricGroup` (little part of complete comment)
"It would be possible to define a class SymmetricGroup as a subclass of
Basic with an argument n, two instances of such a class being equal when their
arguments are equal (SymPy default)."

What led to the thought for this?
---------------------------------
The thought for `__eq__` for groups. Since we (me with @jksuom) are currently
implementing ``FreeGroup``.