invert and mod_inverse modifications
invert is exposed as a function in the * namespace of sympy; it
is also a method of Expr and that method, like cancel, together,
etc..., passes off the arguments directly to the polys functions
which accept gens and args arguments. For that reason, those
arguments were added to invert's argument list.

In addition, Expr.invert passes off numerical arguments to
mod_inverse (moved to reside with igcd, etc..., in core.numbers)
which handles numbers more gracefully (i.e. without the need to
specify generators).