Deep changes in sympy.concrete.guess (see description below).
I rewrote from scratch the find_simple_recurrence_vector function by
using an algorithm inspired by Pade approximants (rather than by using
the PSLQ function from mpmath). The choice came from the fact that the
previous algorithm returned the smallest solution in regards to the
magnitude of coefficients but not the smallest in regards to the order
of recurrence. For that reason, other functions using this result
returned results that were needing to be simplified, which doesn't
seem to be the case any longer.

Furthermore, I added two new types in guess_generating_function:
  * lgf (logarithmic generating function)
  * hlgf (hyperbolic logarithmic generating function).