PermutationGroup() with no argument contains only IdentityElm
Avoid the `ValueError` caused, since PermutationGroup() with no
arguments. One can think of the group generated by an empty
args list as the symmetric group of the empty set, that consists of the
identity element alone.

So set `args = [Permutation()]` and then to go on with no error message.
It should rather return `SymmetricGroup(0)` after PR #10363 is complete.